+++ src/traceback.c
@@ -2,30 +2,15 @@
    why the code was called. */
 
 /*
-  This code was originally  from the Pyrex project:
-  Copyright (C) 2004-2006 Greg Ewing <greg@cosc.canterbury.ac.nz>
+  This code was originally from the Pyrex project:
+  Copyright (C) 2004-2006 Greg Ewing <greg.ewing@canterbury.ac.nz>
 
-  It has been lightly modified to be a part of APSW with permission from Greg.
-  Changes Copyright (C) 2006-2007 Roger Binns <rogerb@rogerbinns.com>
+  It has been lightly modified to be a part of APSW with permission from Greg
+  and to be under the same license (and option for any OSI approved license) as
+  the rest of APSW.
 
-  This software is provided 'as-is', without any express or implied
-  warranty.  In no event will the authors be held liable for any
-  damages arising from the use of this software.
- 
-  Permission is granted to anyone to use this software for any
-  purpose, including commercial applications, and to alter it and
-  redistribute it freely, subject to the following restrictions:
- 
-  1. The origin of this software must not be misrepresented; you must
-     not claim that you wrote the original software. If you use this
-     software in a product, an acknowledgment in the product
-     documentation would be appreciated but is not required.
+  See the accompanying LICENSE file.
 
-  2. Altered source versions must be plainly marked as such, and must
-     not be misrepresented as being the original software.
-
-  3. This notice may not be removed or altered from any source
-     distribution.
   */
 
 /* These are python header files */